.requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
bcrypt==4.0.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
exceptiongroup==1.1.3
flake8==3.9.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
flask-marshmallow==0.14.0
Flask-Migrate==4.0.5
flask-restx==1.1.0
Flask-SQLAlchemy==2.5.1
greenlet==3.0.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.28.1
mccabe==0.6.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==3.11.0
pluggy==1.3.0
psycopg2-binary==2.9.9
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.8.0
pytest==7.4.2
pytest-flake8==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.6
-e /code
six==1.16.0
SQLAlchemy==1.4.0
tomli==2.0.1
typing_extensions==4.8.0
urllib3==2.0.6
Werkzeug==2.2.2
zipp==3.17.0